Refactor for fenix-academic 4.0
