+++ src/main/java/org/fenixedu/academic/json/adapters/ExecutionCourseJsonAdapter.java
@@ -1,4 +1,4 @@
-package org.fenixedu.core.ui.student;
+package org.fenixedu.academic.json.adapters;
 
 import org.fenixedu.academic.domain.ExecutionCourse;
 import org.fenixedu.bennu.core.annotation.DefaultJsonAdapter;
