+++ src/main/java/org/fenixedu/academic/json/adapters/PersonJsonAdapter.java
@@ -1,4 +1,4 @@
-package org.fenixedu.core.ui.student;
+package org.fenixedu.academic.json.adapters;
 
 import org.fenixedu.academic.domain.Person;
 import org.fenixedu.bennu.core.annotation.DefaultJsonAdapter;
