+++ src/main/java/org/fenixedu/bennu/FenixSpringConfiguration.java
@@ -20,10 +20,8 @@ package org.fenixedu.bennu;
 
 import org.fenixedu.bennu.spring.BennuSpringModule;
 
-
-@BennuSpringModule(basePackages = { "org.fenixedu.core.service", "org.fenixedu.academic.ui.spring" },
-        bundles = "FenixEduAcademicResources")
-
+@BennuSpringModule(basePackages = { "org.fenixedu.academic.service", "org.fenixedu.academic.ui.spring" },
+bundles = "FenixEduAcademicResources")
 public class FenixSpringConfiguration {
 
 }
