--- src/main/java/org/fenixedu/core/ui/student/PersonBean.java
@@ -1,45 +1,0 @@
-package org.fenixedu.core.ui.student;
-
-public class PersonBean {
-
-    private String id;
-    private String username;
-    private String name;
-    private String email;
-
-    public PersonBean() {
-    }
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
-
-    public String getUsername() {
-        return username;
-    }
-
-    public void setUsername(String username) {
-        this.username = username;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-
-    public String getEmail() {
-        return email;
-    }
-
-    public void setEmail(String email) {
-        this.email = email;
-    }
-
-}
